api/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu
torchvision==0.16.2+cpu
Pillow>=9.5.0
Flask==3.0.0
Werkzeug==3.0.1
